backend/requirements.txt
psycopg2-binary
flask-sqlalchemy==3.1.1
Flask-Caching==2.3.1
redis==7.4.0
gunicorn==25.3.0
flask-jwt-extended==4.7.1
flask-cors==5.0.1
flask-talisman==1.1.0